Add an --assume-prod option
This is based on --assume-standby-mode. Reduce the option name per suggestion
from Greg (but I kept the original one for standby mode).

The option is only used in check_postgres_checkpoint and allows to confirm or
emit a critical if the server is not in the expected mode.

Note: this can be used in other places, and maybe improved (to reduce the
number og open_controldata calls)

TODO/FIXME:
* I found that --assume-p or --assume-s are viewed by GetOpt like the longer
version of the option, a bug ?
* The original code to call pg_controldata does not work in French (because of
regex/locale). Why not use LANG=C in those checks where there is NO point to
use locale and error prone regex ?