Add `pgagent_jobs` test.
From: "David E. Wheeler" <david@justatheory.com>

This patch adds support for checking for failed pgAgent jobs within a specified
period of time. You can specify either --critical or --warning as a period of
time, and it will report on failures within that period of time previous to the
current time. Job failures are determined by a non-0 status in a job step
record.

Using this test obviously requiers that the pgAgent schema be installed. I've
also included a bunch of unit tests to make sure it works the way I would expect
(the test will create a schema for testing) and documentation.

As part of this, I've introduced the `any_warning` argument to
`validate_range()`. The `pgagent_jobs` test does not care if you specify a
warning value greater than the critical value (indeed, I expect that if one used
both at all, the warning would be much longer). So this new argument prevents
the `range-warnbigtime` or `range-warnbigsize` failures from being triggered.

Cedric: I sorted the POD and added the action_info so that t/05_docs.t is ok.
I also built and push the new .html